Refactor: Move and change loss, norm and error
Norm for each equation are now children of respective Loss classes.
There was a lot of duplicate code - norm is calculated identically as
norm except for boundary conditions. The only thing norm class does
is Loss instance with proper parameters (n_points_error,
integration_rule_norm).

Error package was moved inside loss package - the code is also quite
related.

Functions for creating proper loss, norm and error from Params are
now in single file - `factory.py`. They are short and similar, so it is
more clear now.

Last but not least - x vector is now created directly in loss class and
not passed from the outside. This will allow to integrate in more
clever ways in the future.