Feat: Improve FEM integration
Integration in loss function (and norm) was not precise enough when
using FEM base functions. Which is caused by integrating over
discontinuous function - derivative of FEM base function (pyramid) is
discontinuous.

There was no way to smoothly integrate this change with the code
already written, better integration is too intertwined with loss
function definition. And so for every equation I had to create new
class specifically for loss function with FEM base functions. That
in turn resulted in new Norm classes - they are pretty short but the
code is duplicated :(

The integration itself is working very well. BUT is also super slow.
There is area for improvement - some things can be precomputed
to save time. This was done before outside of loss function definition,
that is in `precompute_base` function. However, it can't be used in bew
integration. Why? Description of new way below.

Basically, every loss function are two linear functions merged together
forming a pyramid, and zeros outside of the pyramid. So we only need to
calculate integral in this small subdomain. In fact in order to make in
more precise we need to calculate it separately for  left and right
sides of pyramid. That is especially true for derivatives, because they
are discontinuous as stated before.

Now, the number of integration points becomes quite useless in this case.
So the new parameters are introduced:
- `n_points_x_fem`
- `n_points_error_fem`
That state number of integration points in each small interval where the
actual integration takes place.

Now precomputed_base that was used before is just list of vectors  for
each base function. And those vectors for FEM are mostly zeroes except
for small interval where they are positive (pyramid). And so it can't be
used for the better integration.